// src/app/+employee/models/employee.ts
export interface Employee {
    employeeId: string;
    name: string;
    firstLastName: string;
    secondLastName: string;
    othersName: string;
    country: string;
    identificationType: string;
    numberIdentification: string;
    email: string;
    admissionDate: Date;
    area: string;
    active: boolean;
}